chatbot-final/chatbot3.1/requirements.txt
fastapi~=0.115.11
uvicorn
pyqt5~=5.15.11
streamlit~=1.37.1
ollama~=0.4.7
requests~=2.32.3
pydantic~=2.10.6
pdfplumber~=0.11.5
pypdf2~=3.0.1
reportlab~=4.3.1
flask~=3.0.3
starlette~=0.46.0
langchain-community~=0.3.19
langchain~=0.3.20
pillow~=10.4.0
wordcloud~=1.9.4
matplotlib~=3.10.1
numpy~=2.2.3